refactor and fixes for model
