+++ app/client-app/src/app/pages/user/Feedback/index.jsx
@@ -18,8 +18,8 @@ import AccountApprovalFeedback from "./Approval";
 import RegisterFeedback from "./Register";
 import PasswordFeedback from "./Password";
 import EmailFeedback from "./Email";
-import { resendEmail } from "api/account";
 import AccountResentComplete from "./Resent";
+import { TokenExpiredFeedback } from "./TokenResend";
 
 export const FeedbackAlert = ({ title, children, ...p }) => (
   <Alert status="info" boxShadow="callout" mt={4} borderRadius={5} p={4} {...p}>
@@ -42,34 +42,8 @@ const LinkError = () => (
 export const DefaultFeedback = () => {
   const { errors } = useQueryStringViewModel();
 
-  const location = window.location;
-  const searchParams = new URLSearchParams(location.search);
-  const pathSegments = location.pathname.split("/");
-
-  let result = {
-    errors: searchParams.get("errors")?.split(",") || [],
-    category: pathSegments[3],
-    state: pathSegments[4],
-    userId: searchParams.get("userId"),
-    code: searchParams.get("code"),
-  };
-
-  const handleResendApprovalEmail = async (userId) => {
-    try {
-      console.log(userId);
-      await resendEmail(userId);
-    } catch (error) {
-      console.error("Failed to resend approval email:", error);
-    }
-  };
-
   // Errors?
-  if (
-    result.category !== "approval" &&
-    result.state !== "tokenexpired" &&
-    errors &&
-    errors.length
-  )
+  if (errors && errors.length)
     return (
       <ErrorsAlert
         title="One or more errors occurred:"
@@ -77,23 +51,6 @@ export const DefaultFeedback = () => {
         boxShadow="callout"
       />
     );
-
-  if (result.category === "approval" && result.state === "tokenexpired") {
-    return (
-      <FeedbackAlert title="Token Expired" status="error">
-        <Box>
-          The approval token has expired. Please request a new approval email.
-        </Box>
-        <Button
-          colorScheme="blue"
-          mt={4}
-          onClick={() => handleResendApprovalEmail(result.userId)}
-        >
-          Resend Approval Email
-        </Button>
-      </FeedbackAlert>
-    );
-  }
 
   // For everything else, there's generic success
   return (
@@ -121,6 +78,7 @@ const UserFeedback = () => (
           <PasswordFeedback path="password/*" />
           <EmailFeedback path="email/*" />
           <AccountResentComplete path="approval/*" />
+          <TokenExpiredFeedback path="approval/tokenexpired/*" />
           <DefaultFeedback default />
         </Router>
       </Flex>
